+++ CHANGES.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+## 2024-02-01 - 0.4.4
+
+- Added the Code component.
+- Removed the NotificationHandler export.
+
 ## 2024-01-31 - 0.4.3
 
-- Fix close button in QueryStackTraceModal
+- Fix close button in QueryStackTraceModal.
 - Export missing StatusLight components and lookups.
 
 ## 2024-01-31 - 0.4.2